Mark xennet.sys as boot-critical to simplify offline installation
Explicitly mark the xennet driver as boot-critical using the
"BootCritical=1" directive.  This causes dism.exe to reflect the
driver when performing an offline installation, thereby making it
available on the next boot (which may be an iSCSI boot, requiring a
functioning network).

This avoids the need to hack the registry to manually install the
driver (by creating the HKLM\CurrentControlSet\Services\xennet key and
corresponding entries in the CriticalDeviceDatabase).

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>