Shim NdisMIndicateReceiveNetBufferLists() to avoid multiple indications
It seems that, on some versions of Windows, something in the network
stack does not work properly with receive indications of more than
one NET_BUFFER_LIST at once.

This patch shims NdisMIndicateReceiveNetBufferLists() to iterate over
a chain of NET_BUFFER_LISTS and indicate them separately. Hopefully
this workaround can be removed in future.

Signed-off-by: Paul Durrant <paul.durrant@citrix.com>