Make sure that a queue of received packets is always pushed
It's possible that the last received packet may suffer an allocation
failure during processing and, as the code stands, any previously
received packets may then not be indicated to the stack.

This patch makes sure that the '!More' condition always results in an
indication, regardless of whether there is an allocation failure.

Signed-off-by: Paul Durrant <paul.durrant@citrix.com>