Update to XENVIF_VIF_INTERFACE version 10
This based on an original patch [1] by Martin Harvey at Citrix. The new Pause flag
added to the XENVIF_RECEIVER_QUEUE_PACKET callback parameters is used when 'low
resources' is indicated to NDIS to request that XENVIF stops queueing packets
for a short period of time (i.e. the time taked to re-schedule the DPC).

[1] https://lists.xenproject.org/archives/html/win-pv-devel/2022-05/msg00044.html

Signed-off-by: Paul Durrant <pdurrant@amazon.com>